replace unit table and implement player actions

+++ src/main/java/io/scherler/games/risk/models/DeploymentResult.java
@@ -9,5 +9,5 @@ public class DeploymentResult {
 
     private String territory;
 
-    private Long totalNumberOfUnits;
+    private Integer totalNumberOfUnits;
 }
